Changed the data type of time to live in hours to Long in ClientTokenSettings

+++ authorization_server_jpa/src/main/java/com/unique/authorization_server_jpa/entity/authorization_server/ClientTokenSettings.java
@@ -27,7 +27,7 @@ public class ClientTokenSettings {
     private Integer id;
 
     @Column(name = "time_to_live_in_hours")
-    private Integer timeToLiveInHours;
+    private Long timeToLiveInHours;
 
     @Column(name = "format")
     private String tokenFormat;
